+++ subprojects/samples/src/main/kotlin/kotools/types/number/StrictlyNegativeDoubleKotlinSample.kt
@@ -4,13 +4,23 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotools.types.experimental.ExperimentalKotoolsTypesApi
 
+@OptIn(ExperimentalKotoolsTypesApi::class)
 internal class StrictlyNegativeDoubleKotlinSample {
-    @OptIn(ExperimentalKotoolsTypesApi::class)
     fun serialization() {
         val number: StrictlyNegativeDouble = StrictlyNegativeDouble.create(-42)
         val encoded: String = Json.encodeToString(number)
         println(encoded) // -42.0
         val decoded: StrictlyNegativeDouble = Json.decodeFromString(encoded)
         println(decoded == number) // true
+    } // END
+
+    fun equalsOverride() {
+        val number: Number = -23
+        val first: StrictlyNegativeDouble =
+            StrictlyNegativeDouble.create(number) // TABS: 1
+        val second: StrictlyNegativeDouble =
+            StrictlyNegativeDouble.create(number) // TABS: 1
+        val result: Boolean = first == second // or first.equals(second)
+        println(result) // true
     } // END
 }